pyproject.toml
[project]
name = "huckleberry-vis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=7.0.1",
    "pandas>=2.3.3",
    "pip>=25.2",
    "plotly>=6.3.1",
    "python-dotenv>=1.1.1",
    "seaborn>=0.13.2",
    "dvc>=3.5.0",
    "pre-commit>=3.4.0",
    "ruff>=0.14.1",
]
